Refactor "ShapefileToGeoJSONService", extricate RGeo usage into FileConverion Service
